+++ models/ObsidianRAGSettings.ts
@@ -1,7 +1,7 @@
-export interface MindMatrixSettings {
+export interface ObsidianRAGSettings {
     vaultId: string;
     excludedFolders: string[];
     excludedFileTypes: string[];
     excludedFilePrefixes: string[];
     excludedFiles: string[];
-} 
+}